+++ docs/content/en/sharding/read-write-separation.md
@@ -1,20 +1,26 @@
-# 分表
+# Read-Write Separation
 
-当某个业务表数据量规模比较大而影响查询性能时，可以采取分表的方式将数据拆分开，`sharding-core` 支持按照`时间`、`取模`等
+As the number of system users increases, the query pressure on the system grows. Using a read-write separation solution can effectively alleviate database query pressure.
 
-## 配置分表
+## Prerequisites
 
-1. 添加包`NetCorePal.Extensions.ShardingCore`引用：
+`sharding-core` does not implement data synchronization to read-only databases. To adopt a read-write separation solution, you need to establish a database data synchronization mechanism, such as `MySQL`'s `binlog` synchronization or `MSSQL`'s `Always On` feature.
+
+## Configuring Read-Write Separation
+
+To use `sharding-core`, you need to modify the registration method of `DbContext`.
+
+1. Add the `NetCorePal.Extensions.ShardingCore` package:
 
       ```shell
       dotnet add package NetCorePal.Extensions.ShardingCore
       ```
-      或者 PackageReference
+      Or use PackageReference:
       ```
       <PackageReference Include="NetCorePal.Extensions.ShardingCore" />
       ```
 
-2. 创建`ApplicationDbContextCreator`
+2. Create `ApplicationDbContextCreator`:
 
     ```csharp
     public class ApplicationDbContextCreator(IShardingProvider provider)
@@ -39,52 +45,22 @@
             return shardingProvider.GetRequiredService<ApplicationDbContext>();
         }
     }
-    
     ```
 
-3. 移除 `AddDbContext` 注册方式
-    ```chsarp
+3. Remove the `AddDbContext` registration method:
+    ```csharp
     builder.Services.AddDbContext<ApplicationDbContext>(options =>
         {
             options.UseMySql(builder.Configuration.GetConnectionString("Mysql"),
                 new MySqlServerVersion(new Version(8, 0, 34)),
                 b => { b.MigrationsAssembly(typeof(Program).Assembly.FullName); });
         });
-    
     ```
-4. 为表添加分表配置：
 
-      ```csharp
-      public class OrderVirtualTableRoute : AbstractSimpleShardingMonthKeyDateTimeVirtualTableRoute<Order>
-      {
-          public override DateTime GetBeginTime()
-          {
-              return DateTime.Now.AddMonths(-3);
-          }
-   
-          public override void Configure(EntityMetadataTableBuilder<Order> builder)
-          {
-              builder.ShardingProperty(o => o.CreationTime);
-          }
-   
-          public override bool AutoCreateTableByTime()
-          {
-              return true;
-          }
-      }
-      ```
-
-
-
-
-5. 使用 `AddShardingDbContext` 注册`ApplicationDbContext`:
+4. Use `AddShardingDbContext` for registration. The following configuration sets up two read-only databases for the data source named `ds0`:
 
     ```csharp
     builder.Services.AddShardingDbContext<ApplicationDbContext>()
-            .UseRouteConfig(op =>
-            {
-               op.AddShardingTableRoute<OrderVirtualTableRoute>(); //注册分表配置
-            })
             .UseConfig(op =>
             {
                 op.ThrowIfQueryRouteNotMatch = true;
@@ -98,14 +74,26 @@
                     builder.UseMySql(con,
                         new MySqlServerVersion(new Version(8, 0, 34)));
                 });
-                op.AddDefaultDataSource("ds0", builder.Configuration.GetConnectionString("Mysql")); //配置写库
+                op.AddDefaultDataSource("ds0", builder.Configuration.GetConnectionString("Mysql")); // Configure write database
+                op.AddReadWriteSeparation(sp => new Dictionary<string, IEnumerable<string>>    // Configure read-only databases
+                    {
+                        {
+                            "ds0",
+                            [builder.Configuration.GetConnectionString("Mysql-Read1"), builder.Configuration.GetConnectionString("Mysql-Read2")]
+                        }
+                    },
+                    readStrategyEnum: ReadStrategyEnum.Loop,
+                    defaultEnableBehavior: ReadWriteDefaultEnableBehavior.DefaultDisable,
+                    defaultPriority: 10,
+                    readConnStringGetStrategy: ReadConnStringGetStrategyEnum.LatestFirstTime
+                );
             })
             .ReplaceService<IDbContextCreator, ApplicationDbContextCreator>()
             .AddShardingCore();
     ```
 
+## Reference Documentation
 
-## 高级
+For more configurations and usage, refer to:
 
-更多关于分表的配置和使用可以参考官方文档： https://xuejmnet.github.io/sharding-core-doc/sharding-table/init/
-
+[https://xuejmnet.github.io/sharding-core-doc/read-write/configure](https://xuejmnet.github.io/sharding-core-doc/read-write/configure)
